Rename MapEntryMapper.Config to MapEntryMappers, and make it a top-level class
