+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/config/internal/ValueColumnImpl.java
@@ -17,7 +17,7 @@ import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 
 import org.jdbi.v3.core.config.ConfigRegistry;
-import org.jdbi.v3.core.mapper.MapEntryMapper;
+import org.jdbi.v3.core.mapper.MapEntryMappers;
 import org.jdbi.v3.sqlobject.config.Configurer;
 import org.jdbi.v3.sqlobject.config.ValueColumn;
 
@@ -25,6 +25,6 @@ public class ValueColumnImpl implements Configurer {
     @Override
     public void configureForMethod(ConfigRegistry registry, Annotation annotation, Class<?> sqlObjectType, Method method) {
         ValueColumn valueColumn = (ValueColumn) annotation;
-        registry.get(MapEntryMapper.Config.class).setValueColumn(valueColumn.value());
+        registry.get(MapEntryMappers.class).setValueColumn(valueColumn.value());
     }
 }
